Changes based on comments completed
I have made the requested changes.

I am wondering if anyone has some suggestions about the MACOSX portion.  I went with Terminal for the text because on the machines I was testing on, it wasn't the text editor that Came up as the first option in spotlight on every machine when I used "TextEdit".  The terminal more consistently launched and since we aren't typing real commands, it sort of works.  Also the lock desktop command works much better from the terminal than is does from spotlight.